agent_finance/requirements.txt
fastapi
psycopg2-binary
python-dotenv
python-jose[cryptography]
cryptography
sqlalchemy
uvicorn
google-generativeai
email-validator
pandas
openpyxl
PyPDF2
python-multipart
requests